+++ forms_pro/tests/test_submission_validation.py
@@ -5,7 +5,11 @@ import json
 import unittest
 from types import SimpleNamespace
 
-from forms_pro.api.submission import _coerce_field_value, _evaluate_conditions, _validate_form_response
+from forms_pro.api.submission.endpoints import (
+    _coerce_field_value,
+    _evaluate_conditions,
+    _validate_form_response,
+)
 from forms_pro.forms_pro.doctype.form_field.form_field import _DISPLAY_ONLY_FIELDTYPES
 from forms_pro.utils.constants import FORMS_PRO_SYSTEM_FIELDNAMES, UNSUPPORTED_FRAPPE_FIELDTYPES
 from forms_pro.utils.form_generator import LINKED_FORM_FIELDOPTIONS, SUBMISSION_STATUS_FIELDOPTIONS
